Window scrolling on vertical movement
Changed the way we specify where the window vs file position. Instead of
specifying where the file is in the window "space", we do the inverse.
Seems more intuitive this way.

Also now only redraw the screen if we've scrolled. That was primarily to
fix the annoying flicker I was seeing. Will probably need something more
capable than that when we get to editing text.

Also added another flake8 ignore rule to play nice with black.