Bugfix: Fix insert() on empty file and delete() handling
Fixing a couple issues with the insert() / delete() handling.

insert() issue:
When opening an empty file and attempting the call to insert(), an `IndexError` exception is raised because the Buffer().lines is an empty list.
```
Traceback (most recent call last):
  File "/home/user/editor.py", line 187, in <module>
    curses.wrapper(main)
  File "/home/user/.pyenv/versions/3.11.3/lib/python3.11/curses/__init__.py", line 94, in wrapper
    return func(stdscr, *args, **kwds)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/user/editor.py", line 181, in main
    buffer.insert(cursor, k)
  File "/home/user/editor.py", line 22, in insert
    current = self.lines.pop(row)
              ^^^^^^^^^^^^^^^^^^^
IndexError: pop from empty list
```

Adding a `try: ... except IndexError: ...` logic corrects this behavior to use an empty string if none is found.

delete(): issue
When hitting the backspace key when the next line is shorter than the current line, or when hitting the backspace key on the last line of the file, the current logic is using the length of the next line to determine if the line should be combined with the next line or not. If on the last line, this causes an `IndexError` exception. If not on the last line, but the cursor is on a Cursor.col is greater than the length of the next line, it falls into the `else` logic combining the next line with the current line incorrectly.

This fix updates it so we are comparing the length of the current line to see if the current line should be combined or not.